Add full op for Arm backend
Implements  the full op which creates a tensor
of a given shape filled with a given value.
The shape and value are set at compile time,
i.e. can't be set by a tensor input.

Refactors tosa_quant_utils.is_quant_node
to handle nodes with no inputs (or outputs)

Does not add a full quantizer annotator, the op
needs to be quantized by a SharedQuantizationSpec

Change-Id: I1cebd1da1af5b9aa726a363431ffc30d8259a0ff
Signed-off-by: Erik Lundell <erik.lundell@arm.com>